Add &nbsp; hack for paragraph spacing
Cube Cobra has changed its Markdown formatting so that paragraphs are
not spaced out like they used to be. For now I'm adding a non-breaking
space between each paragraph, which gets us close to what the output
used to be like. Hopefully they'll revert the change and I'll be able to
revert this hack.